Changed value types from int to float

+++ craftersmine.Riot.Api.League.Challenges/LeagueChallengeThreshholds.cs
@@ -14,36 +14,36 @@ namespace craftersmine.Riot.Api.League.Challenges
         /// Gets a minimum value for challenge to have an Iron badge
         /// </summary>
         [JsonProperty("iron")]
-        public int Iron { get; private set; }
+        public float Iron { get; private set; }
         /// <summary>
         /// Gets a minimum value for challenge to have a Bronze badge
         /// </summary>
         [JsonProperty("bronze")]
-        public int Bronze { get; private set; }
+        public float Bronze { get; private set; }
         /// <summary>
         /// Gets a minimum value for challenge to have a Silver badge
         /// </summary>
         [JsonProperty("silver")]
-        public int Silver { get; private set; }
+        public float Silver { get; private set; }
         /// <summary>
         /// Gets a minimum value for challenge to have a Gold badge
         /// </summary>
         [JsonProperty("gold")]
-        public int Gold { get; private set; }
+        public float Gold { get; private set; }
         /// <summary>
         /// Gets a minimum value for challenge to have a Platinum badge
         /// </summary>
         [JsonProperty("platinum")]
-        public int Platinum { get; private set; }
+        public float Platinum { get; private set; }
         /// <summary>
         /// Gets a minimum value for challenge to have a Diamond badge
         /// </summary>
         [JsonProperty("diamond")]
-        public int Diamond { get; private set; }
+        public float Diamond { get; private set; }
         /// <summary>
         /// Gets a minimum value for challenge to have a Master badge
         /// </summary>
         [JsonProperty("master")]
-        public int Master { get; private set; }
+        public float Master { get; private set; }
     }
 }
